Adding in file transfers.
I'm still not happy with how the file transfers work - they block the
main ui thread while the file is being chosen ... then they block the ui
again while the file is being loaded.

Will work on this a bit more, I'll probably add a dedicated file thread
for this.